requirements.txt
Django==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
pillow==11.1.0
python-dotenv==1.0.1
dj-database-url==2.3.0
